Rename constant container to match semantics
