--- vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/IsoDataModelConstants.kt
@@ -1,79 +1,0 @@
-package at.asitplus.wallet.lib.iso
-
-object IsoDataModelConstants {
-
-    const val DIGEST_SHA_256 = "SHA-256"
-
-    const val VERSION_1_0 = "1.0"
-
-    object DataElements {
-
-        const val FAMILY_NAME = "family_name"
-        const val GIVEN_NAME = "given_name"
-        const val BIRTH_DATE = "birth_date"
-        const val ISSUE_DATE = "issue_date"
-        const val EXPIRY_DATE = "expiry_date"
-        const val ISSUING_COUNTRY = "issuing_country"
-        const val ISSUING_AUTHORITY = "issuing_authority"
-        const val DOCUMENT_NUMBER = "document_number"
-        const val PORTRAIT = "portrait"
-        const val DRIVING_PRIVILEGES = "driving_privileges"
-        const val UN_DISTINGUISHING_SIGN = "un_distinguishing_sign"
-        const val ADMINISTRATIVE_NUMBER = "administrative_number"
-        const val SEX = "sex"
-        const val HEIGHT = "height"
-        const val WEIGHT = "weight"
-        const val EYE_COLOUR = "eye_colour"
-        const val HAIR_COLOUR = "hair_colour"
-        const val BIRTH_PLACE = "birth_place"
-        const val RESIDENT_ADDRESS = "resident_address"
-        const val PORTRAIT_CAPTURE_DATE = "portrait_capture_date"
-        const val AGE_IN_YEARS = "age_in_years"
-        const val AGE_BIRTH_YEAR = "age_birth_year"
-        const val AGE_OVER_18 = "age_over_18"
-        const val ISSUING_JURISDICTION = "issuing_jurisdiction"
-        const val NATIONALITY = "nationality"
-        const val RESIDENT_CITY = "resident_city"
-        const val RESIDENT_STATE = "resident_state"
-        const val RESIDENT_POSTAL_CODE = "resident_postal_code"
-        const val RESIDENT_COUNTRY = "resident_country"
-        const val FAMILY_NAME_NATIONAL_CHARACTER = "family_name_national_character"
-        const val GIVEN_NAME_NATIONAL_CHARACTER = "given_name_national_character"
-        const val SIGNATURE_USUAL_MARK = "signature_usual_mark"
-
-        val ALL_ELEMENTS = listOf(
-            FAMILY_NAME,
-            GIVEN_NAME,
-            BIRTH_DATE,
-            ISSUE_DATE,
-            EXPIRY_DATE,
-            ISSUING_COUNTRY,
-            ISSUING_AUTHORITY,
-            DOCUMENT_NUMBER,
-            PORTRAIT,
-            DRIVING_PRIVILEGES,
-            UN_DISTINGUISHING_SIGN,
-            ADMINISTRATIVE_NUMBER,
-            SEX,
-            HEIGHT,
-            WEIGHT,
-            EYE_COLOUR,
-            HAIR_COLOUR,
-            BIRTH_PLACE,
-            RESIDENT_ADDRESS,
-            PORTRAIT_CAPTURE_DATE,
-            AGE_IN_YEARS,
-            AGE_BIRTH_YEAR,
-            AGE_OVER_18,
-            ISSUING_JURISDICTION,
-            NATIONALITY,
-            RESIDENT_CITY,
-            RESIDENT_STATE,
-            RESIDENT_POSTAL_CODE,
-            RESIDENT_COUNTRY,
-            FAMILY_NAME_NATIONAL_CHARACTER,
-            GIVEN_NAME_NATIONAL_CHARACTER,
-            SIGNATURE_USUAL_MARK,
-        )
-    }
-}
