+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/MobileDrivingLicence.kt
@@ -2,38 +2,38 @@
 
 package at.asitplus.wallet.lib.iso
 
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.ADMINISTRATIVE_NUMBER
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.AGE_BIRTH_YEAR
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.AGE_IN_YEARS
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.AGE_OVER_18
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.BIRTH_DATE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.BIRTH_PLACE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.DOCUMENT_NUMBER
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.DRIVING_PRIVILEGES
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.EXPIRY_DATE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.EYE_COLOUR
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.FAMILY_NAME
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.FAMILY_NAME_NATIONAL_CHARACTER
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.GIVEN_NAME
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.GIVEN_NAME_NATIONAL_CHARACTER
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.HAIR_COLOUR
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.HEIGHT
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.ISSUE_DATE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.ISSUING_AUTHORITY
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.ISSUING_COUNTRY
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.ISSUING_JURISDICTION
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.NATIONALITY
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.PORTRAIT
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.PORTRAIT_CAPTURE_DATE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.RESIDENT_ADDRESS
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.RESIDENT_CITY
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.RESIDENT_COUNTRY
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.RESIDENT_POSTAL_CODE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.RESIDENT_STATE
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.SEX
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.SIGNATURE_USUAL_MARK
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.UN_DISTINGUISHING_SIGN
-import at.asitplus.wallet.lib.iso.IsoDataModelConstants.DataElements.WEIGHT
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.ADMINISTRATIVE_NUMBER
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.AGE_BIRTH_YEAR
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.AGE_IN_YEARS
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.AGE_OVER_18
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.BIRTH_DATE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.BIRTH_PLACE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.DOCUMENT_NUMBER
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.DRIVING_PRIVILEGES
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.EXPIRY_DATE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.EYE_COLOUR
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.FAMILY_NAME
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.FAMILY_NAME_NATIONAL_CHARACTER
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.GIVEN_NAME
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.GIVEN_NAME_NATIONAL_CHARACTER
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.HAIR_COLOUR
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.HEIGHT
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.ISSUE_DATE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.ISSUING_AUTHORITY
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.ISSUING_COUNTRY
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.ISSUING_JURISDICTION
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.NATIONALITY
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.PORTRAIT
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.PORTRAIT_CAPTURE_DATE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.RESIDENT_ADDRESS
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.RESIDENT_CITY
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.RESIDENT_COUNTRY
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.RESIDENT_POSTAL_CODE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.RESIDENT_STATE
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.SEX
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.SIGNATURE_USUAL_MARK
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.UN_DISTINGUISHING_SIGN
+import at.asitplus.wallet.lib.iso.MobileDrivingLicenceDataElements.WEIGHT
 import at.asitplus.wallet.lib.jws.ByteArrayBase64UrlSerializer
 import io.github.aakira.napier.Napier
 import io.matthewnelson.encoding.base16.Base16
